In progress commit on issue #3.
The Top Nav bar and the Footer components are in place, and responsive.

A Layout component was created to wrap the nav with the rest of the
layout as the top nav position will change to a side nav on a logged in
desktop view.

A helpers folder was created and currently only holds the isMobile hook.
This hook returns true if the screen width is less than 768px. The
hook also registers and cleans up a window resize event listener.

An Icons component folder was created to organize any svg Icons that
will be used. Currently we have the Close and EllipsisMenu icons.
The icons can take their own classes and accepts an onClick callback.