Updated the Login to take a user to their dashboard which is the Lists
page. Also added a WithAuth component that will redirect a user to the
login page if they are not logged in. I refactored the NavLink logic to
use a provider so that the available link logic was in a central
location. A user provider was also added to store the user information
and serves as a loggedIn feature by referencing the isAuthenticated
property.