Added a GlobalStore to hold our user's list data. This will cut down
the amount of API calls we need to make to the server. By allowing us
to reuse data we have already fetched.

The view logic has been modified so that we can render an endpoint with
an id parameter. This will allow us to render a specific view. In our
case this has been applied to an individual list.

We are able to click on a list from the list of lists and be taken to
the list view. The data is only being logged at the moment. But logic
has been added to use the data from global state or fetch and add it if
it is missing. It will be missing if the route was navigated to directly
from the url rather than in a controlled manner through the application,
since the global state isn't persisted through page refreshes.