Created a Toast Provider and all the components needed to display toast
messages to the user. Toasts are themed as info, success, warning, and
error, and they have the relevant visual theming. Toasts are displayed
in a portal, so they can be displayed anywhere in the app. Toasts are
displayed by calling the useToastMessage hook, which returns a function
that can be called to display a toast. The toast clears after 5 seconds,
on its own or can be cleared by clicking the X button. Currently only
one toast can be displayed at a time, but this can be changed in the
future.