Cleaned up the User provider and moved the session management to its
own file. This makes the User provider more readable and easier to
understand.

Tokens are good for 60 mins - I originally thought I had set it to 15
but I was wrong. Users are validated on every page request, we operate
on a model of zero trust. If a user is inactive for 10 mins they will
automagically be logged out. IF they are actively navigating around,
their session will be refreshed and they will be logged in for another
10 mins or less depending on the status of their JWT  which is only
good for a max of 60 mins. Then they will have to reauthenticate.