Working on a EditableContent component that can be used for any
editable text content.

The nav was adjusted to have a back arrow on mobile.

The List item was adjusted for display issues on smaller screens.

The Editable content hasn't been completed just yet as I am still
pondering how to handle UPC product name updates by the user. I am
thinking that I don't want to allow the user to update any of the UPC
data. We could create a mapping of name changes for each user and use
that to update the name for the user, and then we could use it to update
the name in the database if it meets a certain threshold of users that
have made the same change. This would allow us to keep the database
clean and accurate, while still allowing the user to have a custom
experience. Either way, I want to know how this will ultimately work
before I finish the EditableContent component.