Adjusted to hooks for data fetching over the suspense and the stale
errors are gone. Added the ability for creating new lists through the
app and everything appears to be working as expected. Adjusted the
user session logic to be more robust and to renew the session as long
as the token is valid and the user is active on the app. Max time is
50 minutes.