Added Service Worker for PWA functionality. Corrected some mobile
issues that were causing the app to not work properly on mobile.

The close handlers have been adjusted for the Editable Field components
used in the ListItem and ListCard components. When tapping on the text
to capture it to edit, the component would register a click in the
component's parent and would close. This has been corrected.

On mobile the list card menu can't actually enter a hover state, so the
hover colors were made the static color instead.

For some reason the SVG icons for just the Increase and Decrease buttons
were not rendering on mobile even though they are the same type of icon
used elsewhere. I have replaced them with + and - text for now and
will revisit this later.