Styling and mobile quirks have been mainly adjusted. The check mark
still needs attention on mobile. When testing on the desktop the mobile
emulators do not show the same errors as the actual mobile device. The
check mark isn't styled and isn't correctly positioned but it is a bit
difficult to reproduce in a test environment. For now the following
issues have been addressed.

Touch and mouse issues when trying to highlight the text in the
EditableContent component.

Redirect to previous page if asked to reauthenticate.

Corrected an issue when editing a product name would act as if an error
ocurred but in reality the product name was being updated.

Corrected styling issues on Mobile devices for the increase quantity
button. The + symbol was out of alignment but this doesn't present
itself in a test environment and may not exist beyond apple devices.

Form Action Button was adjusted to say Processing and not reset itself.
The button's state can only be reset by an unmounting of the component
or after a 12 second timeout. Once the processing spinner has gone
through all its color changes the button will reset.