Adjusted some issues with mobile and made other minor adjustments.
Some icons have been updated to use Heroicons React. The SVG icons we
were using were not rendering properly on mobile, even though they were
copied from the Heroicons website.
    Icons adjusted:
        - PlusCircle
        - MinusCircle
        - Check (Wasn't an icon before, but is now)

The decreaseQuantity button has been adjusted so that it must be double
clicked or tapped to decrease the quantity. This is to prevent an
accidental decrease in quantity.

The double menu icon has been adjusted to show/hide the add/remove
quantity buttons on mobile.

The Submit button on the signup/login forms has been adjusted to reset
on error.

The Set Default menu option has been adjusted to stay amber on hover.

Completed status has been added but the functionality is local.
We can offer a status sync for premium users in the future.

The List Item renderer has been adjusted so that the items keep an
index. This prevents the order from changing when item quantities are
changed.

Everything that the MVP needed is now complete and rudimentary tests
show that it works on desktop and mobile.