A demo test version of the camera barcode scanner is being rolled out.
The barcode scanner is quirky and needs improvement. On an unsuccessful
scan, the user is given the option to manually enter the barcode.

This will need to be improved in the future. The camera scanner doesn't
work so well with smaller or thin barcodes. But it does work in some
cases. Ideally this application should be used with the TrashScan Device
which is a dedicated barcode scanner.