Refactoring and adjusting some oddities here and there.
There was an issue when updating a product's name the linked list was
not being updated. This has been corrected.

When deleting a list the list would be opened again. This has also been
corrected.

The rendering logic for an individual list has been cleaned up and the
logic for consolidating and creating the linked list has been moved to
a helper file in the List directory.