Work in progress. Still needs to be tested and cleaned up.
Saving this here for now. But we have implemented swipe actions for
touch devices. This is a first attempt. We are trying to provide three
options but we only have two directions to work with.

Currently the swipe to the left provides an option to delete one item,
a full swipe to the left removes all items and a full swipe to the
right adds one item.