Implemented the swipe actions. They have been slightly modified from
the original. To delete all a button must be pressed.
Sliding left or right will display the add one and delete one/all
buttons. A quick swipe left will delete only one. A quick swipe right
will add one. Additional changes were implemented to adjust for the
issues with updating the linkedList, it wasn't a huge deal but when an
item was deleted it wasn't removed from the linkedList unless the
list was forcefully deleted. Everything at the moment is working
as expected.