Addressed some issues with the list items and corrected an issue
when using a touch screen device when trying to scroll. The smallest
deviation right or left would cause an item that you were scrolling On
to show the add or delete buttons. Now if you are scrolling list items
will not move so you can't accidentally delete or add an item and you
dont have to constantly reset them.

Was passing the wrong ID for some components so now the status displays
and updates correctly, and the swipe actions work for deleting items.

This commit will close issue #15.