Refactoring, removing unnecessary or unused code.
The userProvider which handles all the authentication logic for the user
was modified. Every request in the app is validated
and the previous setup was looking up the JWT token in localStorage
and decoding it every time. This was changed to only decode the token
when it is fetched the first time, it is validated and then stored in
state. The toast and toaster were modified to not use a portal. This
was making the page unresponsive behind the toast message which leads
to a bad user experience. The portal was removed and the toast behaves
as expected. The linkedList and suspender were removed as they were
not being used. The types were modified to have the correct types for
the JWT.