Added an ErrorBoundary component to catch andy uncaught errors in the
application.  This will log an error message to the console and ensure
the app continues to display.

The list item component was adjusted slightly so that the add/remove
buttons do not auto-hide on desktop or non-touch devices.

Adjusted a potential issue with the user provider where the user
information wasn't being updated when the token was removed in the
session manager, this should be happening, but a listener was added
on all localStorage events, and we ensure the token still exists or
handle it accordingly.